IKASAN-2135 final changes for job lock management work.

+++ ikasaneip/spec/service/scheduled/src/main/java/org/ikasan/spec/scheduled/context/model/JobLock.java
@@ -27,13 +27,13 @@ public interface JobLock extends Serializable {
      *
      * @param lockCount
      */
-    void setLockCount(long lockCount);
+    void setLockCount(int lockCount);
 
     /**
      * Get the name of the lock count.
      * @return
      */
-    long getLockCount();
+    int getLockCount();
 
     /**
      * Set the scheduler jobs.
